+++ discord/objects/components/__init__.py
@@ -2,5 +2,6 @@ from .checkbox import CheckboxComponent
 from .checkbox_group import CheckboxGroupComponent
 from .checkbox_group_option import CheckboxGroupOption
 from .file_upload import FileUploadComponent
+from .label import LabelComponent
 from .radio_group import RadioGroupComponent
 from .radio_group_option import RadioGroupOption