+++ discord/objects/components/_base.py
@@ -1,87 +1,123 @@
 from ..._dataclass import dataclass
 from ...enums import ComponentType
 from ...utils import MISSING, Optional
+from typing import Any
 
 
 @dataclass
-class Component:
+class Component[T]:
   id: Optional[int]
   """32-bit integer used as an optional identifier for component."""
   
   type: ComponentType
   """The type of the component."""
   
-  def __class_getitem__(cls, component_type: int) -> Optional[type]:
+  def __new__(cls, **data: dict[str, Any]) -> Optional[T]:
     """:meta private:"""
-    if not isinstance(component_type, int):
-      raise TypeError(f"component_type: Must be an instance of {int}; not {component_type.__class__}")
-    if component_type < 0:
-      raise ValueError(f"component_type: Must be greater than or equal to 0")
-    match component_type:
+    match data["type"]:
       case ComponentType.ACTION_ROW:
         from .action_row import ActionRowComponent
-        return ActionRowComponent
+        component_cls: T = ActionRowComponent
       case ComponentType.BUTTON:
         from .button import ButtonComponent
-        return ButtonComponent
+        component_cls: T = ButtonComponent
       case ComponentType.STRING_SELECT:
         from .string_select import StringSelectComponent
-        return StringSelectComponent
+        component_cls: T = StringSelectComponent
       case ComponentType.TEXT_INPUT:
         from .text_input import TextInputComponent
-        return TextInputComponent
+        component_cls: T = TextInputComponent
       case ComponentType.USER_SELECT:
         from .user_select import UserSelectComponent
-        return UserSelectComponent
+        component_cls: T = UserSelectComponent
       case ComponentType.ROLE_SELECT:
         from .role_select import RoleSelectComponent
-        return RoleSelectComponent
+        component_cls: T = RoleSelectComponent
       case ComponentType.MENTIONABLE_SELECT:
         from .mentionable_select import MentionableSelectComponent
-        return MentionableSelectComponent
+        component_cls: T = MentionableSelectComponent
       case ComponentType.CHANNEL_SELECT:
         from .channel_select import ChannelSelectComponent
-        return ChannelSelectComponent
+        component_cls: T = ChannelSelectComponent
       case ComponentType.SECTION:
         from .section import SectionComponent
-        return SectionComponent
+        component_cls: T = SectionComponent
       case ComponentType.TEXT_DISPLAY:
         from .text_display import TextDisplayComponent
-        return TextDisplayComponent
+        component_cls: T = TextDisplayComponent
       case ComponentType.THUMBNAIL:
         from .thumbnail import ThumbnailComponent
-        return ThumbnailComponent
+        component_cls: T = ThumbnailComponent
       case ComponentType.MEDIA_GALLERY:
         from .media_gallery import MediaGalleryComponent
-        return MediaGalleryComponent
+        component_cls: T = MediaGalleryComponent
       case ComponentType.FILE:
         from .file import FileComponent
-        return FileComponent
+        component_cls: T = FileComponent
       case ComponentType.SEPARATOR:
         from .separator import SeparatorComponent
-        return SeparatorComponent
+        component_cls: T = SeparatorComponent
       case ComponentType.CONTAINER:
         from .container import ContainerComponent
-        return ContainerComponent
+        component_cls: T = ContainerComponent
       case ComponentType.LABEL:
         from .label import LabelComponent
-        return LabelComponent
+        component_cls: T = LabelComponent
       case ComponentType.FILE_UPLOAD:
         from .file_upload import FileUploadComponent
-        return FileUploadComponent
+        component_cls: T = FileUploadComponent
       case ComponentType.RADIO_GROUP:
         from .radio_group import RadioGroupComponent
-        return RadioGroupComponent
+        component_cls: T = RadioGroupComponent
       case ComponentType.CHECKBOX_GROUP:
         from .checkbox_group import CheckboxGroupComponent
-        return CheckboxGroupComponent
+        component_cls: T = CheckboxGroupComponent
       case ComponentType.CHECKBOX:
         from .checkbox import CheckboxComponent
-        return CheckboxComponent
-    return MISSING
+        component_cls: T = CheckboxComponent
+    component_cls: Optional[T] = MISSING
+    return super().__new__(component_cls) if component_cls else MISSING
 
 
 @dataclass
 class InteractiveComponent:
   custom_id: str
   """Developer-defined identifier (1-100 characters)."""
+
+
+class LabelChildComponent[T]:
+  def __new__(cls, **data: dict[str, Any]) -> Optional[T]:
+    """:meta private:"""
+    match data["type"]:
+      case ComponentType.TEXT_INPUT:
+        from .text_input import TextInputComponent
+        component_cls: T = TextInputComponent
+      case ComponentType.STRING_SELECT:
+        from .string_select import StringSelectComponent
+        component_cls: T = StringSelectComponent
+      case ComponentType.USER_SELECT:
+        from .user_select import UserSelectComponent
+        component_cls: T = UserSelectComponent
+      case ComponentType.ROLE_SELECT:
+        from .role_select import RoleSelectComponent
+        component_cls: T = RoleSelectComponent
+      case ComponentType.MENTIONABLE_SELECT:
+        from .mentionable_select import MentionableSelectComponent
+        component_cls: T = MentionableSelectComponent
+      case ComponentType.CHANNEL_SELECT:
+        from .channel_select import ChannelSelectComponent
+        component_cls: T = ChannelSelectComponent
+      case ComponentType.FILE_UPLOAD:
+        from .file_upload import FileUploadComponent
+        component_cls: T = FileUploadComponent
+      case ComponentType.RADIO_GROUP:
+        from .radio_group import RadioGroupComponent
+        component_cls: T = RadioGroupComponent
+      case ComponentType.CHECKBOX_GROUP:
+        from .checkbox_group import CheckboxGroupComponent
+        component_cls: T = CheckboxGroupComponent
+      case ComponentType.CHECKBOX:
+        from .checkbox import CheckboxComponent
+        component_cls: T = CheckboxComponent
+    component_cls: Optional[T] = MISSING
+    return super().__new__(component_cls) if component_cls else MISSING